fix imports in openvino cli command
